prplmesh.init: set ebtables to not bridge 1905.1 frames
1905.1 frames must be handled per-interface by the prplMesh daemon, so
they must not be forwarded over any bridge. Also, prplMesh will only
listen to the individual interfaces, not to the bridge, so the frames
must also not be forwarded up to the bridge interface.

Therefore, block the forwarding of the 1905.1 frames from the BROUTING
ebtable. This causes them to arrive on the individual interface instead
of on the bridge.